+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Added `alwaysLeafCategoryAuction` settings parameter to run auctions only over leaf categories
+- Added `activateDebugSponsoredTags` settings parameter to display the (Ad) tag next to sponsored products
+
 ## [2.1.1] - 2025-03-11
 
 ### Added